chore(all): bumped releases to official
