+++ Cargo.toml
@@ -3,8 +3,8 @@
 resolver = "3"
 # Include both parts of the library in the workspace
 members = [
-    "bora",
     "deboa",
+    "deboa-bora",
     "deboa-extras",
     "deboa-macros",
     "examples/blog",
