+++ deboa-bora/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "bora"
+name = "deboa-bora"
 edition = "2021"
-version = "0.0.1-beta.3"
+version = "0.0.2"
 authors = ["Rogerio Araújo <rogerio.araujo@gmail.com>"]
 description = "bora macro for deboa"
 license = "MIT"
@@ -22,9 +22,9 @@ xml = ["deboa-extras/xml"]
 msgpack = ["deboa-extras/msgpack"]
 
 [dependencies]
-deboa = { path = "../deboa", version = "0.0.5-beta.3", features = ["http1"] }
-vamo = { path = "../vamo", version = "0.0.1-beta.3" }
-deboa-extras = { path = "../deboa-extras", version = "0.0.1-beta.3", features = ["json"] }
+deboa = { path = "../deboa", version = "0.0.5", features = ["http1"] }
+vamo = { path = "../vamo", version = "0.0.1" }
+deboa-extras = { path = "../deboa-extras", version = "0.0.1", features = ["json"] }
 proc-macro2 = "1.0.101"
 tokio = { version = "1.47.1", features = ["full"] }
 serde = "1.0.219"
