+++ deboa-bora/tests/bora/post.rs
@@ -1,11 +1,12 @@
-use bora::bora;
 use deboa::errors::DeboaError;
+use deboa_bora::bora;
 use deboa_tests::utils::JSONPLACEHOLDER;
 use serde::{Deserialize, Serialize};
 use vamo::Vamo;
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Post {
+    pub id: u32,
     pub title: String,
     pub body: String,
     #[serde(rename = "userId")]
@@ -14,26 +15,24 @@ pub struct Post {
 
 #[bora(
       api(
-        patch(name="patch_post", path="/posts/<id:i32>", req_body=Post, format="json"),
+        post(name="create_post", path="/posts", req_body=Post, format="json"),
       )
     )]
 pub struct PostService;
 
 #[tokio::test]
-async fn test_patch_by_id() -> Result<(), DeboaError> {
+async fn test_get_by_id() -> Result<(), DeboaError> {
     let client = Vamo::new(JSONPLACEHOLDER)?;
 
     let mut post_service = PostService::new(client);
 
     post_service
-        .patch_post(
-            1,
-            Post {
-                title: "title".to_string(),
-                body: "body".to_string(),
-                user_id: 1,
-            },
-        )
+        .create_post(Post {
+            id: 1,
+            title: "title".to_string(),
+            body: "body".to_string(),
+            user_id: 1,
+        })
         .await?;
     Ok(())
 }
