+++ deboa-extras/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "deboa-extras"
-version = "0.0.1-beta.3"
+version = "0.0.1"
 edition = "2021"
 authors = ["Rogerio Araújo <rogerio.araujo@gmail.com>"]
 repository = "https://github.com/ararog/deboa"
@@ -13,7 +13,7 @@ publish = true
 rust-version = "1.64.0"
 
 [dependencies]
-deboa = { path = "../deboa", version = "0.0.5-beta.3", features = ["http1"], optional = true }
+deboa = { path = "../deboa", version = "0.0.5", features = ["http1"], optional = true }
 rmp-serde = { version = "1.3.0", optional = true }
 serde = { version = "1.0.217", features = ["derive"], optional = true }
 serde-xml-rust = { version = "0.6.0", optional = true }
