+++ deboa-macros/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "deboa-macros"
 edition = "2021"
-version = "0.0.1-beta.3"
+version = "0.0.1"
 authors = ["Rogerio Araújo <rogerio.araujo@gmail.com>"]
 description = "macros for deboa"
 license = "MIT"
@@ -20,9 +20,9 @@ xml = ["deboa-extras/xml"]
 msgpack = ["deboa-extras/msgpack"]
 
 [dependencies]
-bora = { path = "../bora", version = "0.0.1-beta.3" }
-deboa = { path = "../deboa", version = "0.0.5-beta.3", features = ["http1"] }
-deboa-extras = { path = "../deboa-extras", version = "0.0.1-beta.3", features = ["json"] }
+deboa = { path = "../deboa", version = "0.0.5", features = ["http1"] }
+deboa-bora = { path = "../deboa-bora", version = "0.0.2" }
+deboa-extras = { path = "../deboa-extras", version = "0.0.1", features = ["json"] }
 tokio = { version = "1.47.1", features = ["full"] }
 serde = "1.0.219"
