+++ deboa-macros/src/lib.rs
@@ -1,4 +1,4 @@
-pub use bora::bora;
+pub use deboa_bora::bora;
 
 #[macro_export]
 /// Make a GET request to the specified URL.
