+++ deboa/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "deboa"
-version = "0.0.5-beta.3"
+version = "0.0.5"
 edition = "2021"
 authors = ["Rogerio Araújo <rogerio.araujo@gmail.com>"]
 repository = "https://github.com/ararog/deboa"
