+++ vamo-macros/Cargo.toml
@@ -16,9 +16,9 @@ rust-version = "1.64.0"
 proc-macro = true
 
 [dependencies]
-deboa = { path = "../deboa", version = "0.0.5-beta.3", features = ["http1"] }
-deboa-extras = { path = "../deboa-extras", version = "0.0.1-beta.3", features=["json"] }
-vamo = { path = "../vamo", version = "0.0.1-beta.3" }
+deboa = { path = "../deboa", version = "0.0.5", features = ["http1"] }
+deboa-extras = { path = "../deboa-extras", version = "0.0.1", features=["json"] }
+vamo = { path = "../vamo", version = "0.0.1" }
 proc-macro2 = "1.0.101"
 tokio = { version = "1.47.1", features = ["full"] }
 serde = "1.0.219"
