+++ vamo/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "vamo"
-version = "0.0.1-beta.3"
+version = "0.0.1"
 edition = "2021"
 authors = ["Rogerio Araújo <rogerio.araujo@gmail.com>"]
 repository = "https://github.com/ararog/deboa"
@@ -13,8 +13,8 @@ publish = true
 rust-version = "1.64.0"
 
 [dependencies]
-deboa = { path = "../deboa", version = "0.0.5-beta.3", features = ["http1"] }
-deboa-extras = { path = "../deboa-extras", version = "0.0.1-beta.3", features = ["json"] }
+deboa = { path = "../deboa", version = "0.0.5", features = ["http1"] }
+deboa-extras = { path = "../deboa-extras", version = "0.0.1", features = ["json"] }
 http = "1.3.1"
 serde = "1.0.228"
 url = "2.5.7"
